🚧  commandUtils.jsを削除し、TypeScriptでcommandUtils.tsを新規作成 ## TODO - 型のエラー修正を忘れてる
